gitk: Speed up the reading of references
We were doing two execs for each tag - one to map the tag ID to a commit
ID (which git ls-remote was already giving us as refs/tags/foo^{}) and
one to read the contents of the tag for later display.  This speeds up
the process by not reading the contents of the tag (instead it is read
later if needed), using the refs/tags/foo^{} lines from git ls-remote,
and using string operations instead of regexps.

Signed-off-by: Paul Mackerras <paulus@samba.org>