gitk: Use non-topological order for the date mode
This adds code for detecting and coping with misorderings (i.e. commits
that come after one or more of their parents) in the git-rev-list
output.  This means that for date mode (invoked by the -d flag to gitk)
we don't need to use the --topo-order or --date-order flags to
git rev-list.  Instead we have a reorder buffer to cope with minor
misorderings.  If we get a misordering that the reorder buffer can't
cope with, we freeze the layout and display at that point, reorder the
commits as we get them, and then re-do the layout from scratch when
we have read all the commits.

The advantage of this is faster startup, particularly in the cold-cache
case or with large unpacked repositories.  Without --topo-order or
--date-order, git rev-list doesn't have to read in all of the commits
before it can produce some output and we can start drawing stuff.

This also gets rid of the childlist variable; the information in there
was a duplicate of the children array.

Signed-off-by: Paul Mackerras <paulus@samba.org>