gitk: Improve the behaviour of the initial selection
It used to be that if you clicked on a line while gitk was still drawing
stuff, it would immediately re-select the first line of the display.
This fixes that, and also defaults to selecting the second line if the
first line is the fake commit for the local changes.

Signed-off-by: Paul Mackerras <paulus@samba.org>