gitk: Only draw the graph lines for the visible parts
This changes the way the graph lines are drawn so that we just draw
lines between the graph rows that are drawn, rather than drawing the
whole length of a segment of the graph line (from endpoint or arrow
to endpoint or arrow).

This means that we don't have to optimize all the rows from the top
down; instead we can just optimize the rows we are drawing plus some
extra on either side.  (Currently we don't actually do enough for
some pathological situations.)

This also adds a firstchild($view,$id) array which tells us which row
the line for a commit terminates on at the top end.

Signed-off-by: Paul Mackerras <paulus@samba.org>