gitk: Store ids in rowrangelist and idrowranges rather than row numbers
This removes the need for insertrow and removerow to go through
rowrangelist and idrowranges and adjust a lot of entries.  The first
entry for a given id is now the row number of the first child, not
that row number + 1, and rowranges compensates for that so its callers
didn't have to change.  This adds a ranges argument to drawlineseg
so that we can avoid calling rowranges a second time inside drawlineseg
(all its callers already called rowranges).

Signed-off-by: Paul Mackerras <paulus@samba.org>