gitk: Further tweaks to the graph-line drawing algorithm
This gets rid of the use of rowoffsets in drawlineseg et al., and
dispenses with the firstchild function.  Instead of using rowoffsets,
drawlineseg now records which column the line is in for each row it
is looking at and uses that (it was doing the search already but not
saving the results).  Instead of using firstchild, we now test if
the parent has a line on the same row that the child appears; if so
the child is not the first child of that parent.

Signed-off-by: Paul Mackerras <paulus@samba.org>