gitk: Use tokens to map ids to row numbers
So that rows can be inserted easily without having to update a lot
of entries of the commitrow array, this changes gitk to map ids to
row numbers via a 'token', which is a list of integers.  Tokens have
the obvious ordering, and the tokens are assigned to the ids in the
graph in increasing order going down the list (increasing row number).
This means that a binary search on the list of token values can be used
to map a token value to a row number.  It is possible to assign a new
token value between any two existing token values, by adding a new
element to the list if necessary.

This gets rid of the commitrow array entirely and uses a new rowofcommit
function instead.  This also changes parentlist back to a parents array
(so that one can get the parents of an id without having to map the id to
a row number), removes the rowrangelist list and just uses the
idrowranges array instead (but with the view as part of the index),
and changes is_ancestor and is_descendent to do comparisons on tokens
rather than row numbers.

Signed-off-by: Paul Mackerras <paulus@samba.org>