gitk: Use thread shared variables for several variables
This converts displayorder, commitlisted, vtokenlist, committok,
parents, children, commitdata, commitinfo, desc_heads, and rowidlist
to thread-shared variables (TSVs).

rowidlist is both a normal variable used by layoutrows etc., and a
TSV.  Stuff is transferred from the normal variable to the TSV when
it is ready to be displayed.

This also changes findcrossings to not use rowoffsets, so that it
doesn't need to be made a TSV.

Signed-off-by: Paul Mackerras <paulus@samba.org>