gitk: Make use of the Tcl thread extension if available
This loads the Tcl thread extension if it is available, and uses it to
run the graph loading and layout in one thread, and the computation of
ancestor/descendent tags and descendent heads in another thread.  This
means that the main thread just has to do the GUI, and can remain
responsive while the graph computations are being done.  The processing
of search operations and previous/next highlighted commit operations
are still done in the main thread, but should probably also be
offloaded to another thread.

This uses the tsv (thread shared variable) commands to communicate
between the threads.  If the thread extension is not available, then
everything runs in the main thread as before, and we have a set of fake
tsv commands that just manipulate global variables.  However, there are
some places where calling the fake tsv commands hurts performance too
much, so I have arranged for some preprocessing of the layout code so
that it turns into code using the tsv commands when we're running
threaded, or simple set, lappend, info exists etc. commands when running
non-threaded.  These places are marked by words between percent signs,
such as %iflthread%, %tsvexists% etc.  See the preproc command for how
these are processed.

Signed-off-by: Paul Mackerras <paulus@samba.org>