gitk: Fix bugs in line drawing and reload
First, the arrows-on-diagonal-lines workaround was getting activated
when it shouldn't have been, resulting in the graph lines sometimes
looking curved.  Secondly, the vrowidlist array wasn't getting reset
when we did a reload, resulting in a completely messed-up graph if
the order of commits was different after the reload.  Also the
desc_heads, desc_tags and anc_tags arrays weren't getting reset either.

Signed-off-by: Paul Mackerras <paulus@samba.org>