gitk: Improve performance of threaded mode on single-cpu systems
On single-cpu systems, with threads enabled, the whole thing was taking a
lot longer than with threads disabled.  The reason was that each call to
getcommitlines, layoutmore, getallclines etc. were only processing a few
commits each time, rather than having the work batched up to reduce
overheads.

The best fix turns out to be to set the pipes from the git rev-list
processes to blocking mode.  That batches up the work nicely and reduces
the time taken by about 40%, to not much more than the non-threaded case
takes.

Signed-off-by: Paul Mackerras <paulus@samba.org>